requirements.txt
fastapi
uvicorn
pandas
scikit-learn
joblib
sqlalchemy[asyncio]
psycopg[binary]
python-dotenv
